requirements.txt
requests==2.11.1
Pillow==3.3.1
